Move shard distribution unit tests into test_moe_compute_6U.py
Consolidate the 13 CPU-only shard formula tests from the standalone
test_moe_compute_utils_shape.py into the main MoE test suite so all
MoE tests live in one place.
